pyproject.toml
[tool.poetry]
name = "k3d-cluster"
version = "0.1.0"
description = ""
authors = ["smaddanki"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
kafka-python-ng = "^2.2.3"
six = "^1.16.0"



[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
